TerminalExecutionConsole instead of ScrollPane & Label [RIP-59]
